[#43] Add "query_image" service
