+++ custom_components/extended_openai_conversation/const.py
@@ -84,3 +84,5 @@ DEFAULT_CONF_FUNCTIONS = [
 ]
 CONF_ATTACH_USERNAME = "attach_username"
 DEFAULT_ATTACH_USERNAME = False
+
+SERVICE_QUERY_IMAGE = "query_image"
